+++ src/core/control/id_control_msg.rs
@@ -1,8 +1,11 @@
 use serde::Deserialize;
 
 use crate::core::{
-    Msg, Result, UampApp, config::ConfigMsg, library::SongId,
-    player::Playlist, server::SubMsg,
+    Error, Msg, Result, UampApp,
+    config::ConfigMsg,
+    library::SongId,
+    player::Playlist,
+    server::{SubMsg, sub::InsertIntoPlaylist},
 };
 
 #[derive(Debug, Deserialize)]
@@ -16,6 +19,13 @@ pub enum IdControlMsg {
         songs: Vec<SongId>,
         position: usize,
         play: bool,
+    },
+    InsertIntoPlaylist {
+        songs: Vec<SongId>,
+        /// Position at which to insert.
+        position: usize,
+        /// Playlist index in playlist stack.
+        playlist: usize,
     },
     SetConfig(serde_json::Value),
 }
@@ -55,6 +65,28 @@ impl UampApp {
                 self.client_update_set_playlist(|a| {
                     SubMsg::PushPlaylist(a.into())
                 });
+            }
+            IdControlMsg::InsertIntoPlaylist {
+                mut songs,
+                position,
+                playlist,
+            } => {
+                let Some(plist) = self.player.get_playlist_mut(playlist)
+                else {
+                    return Error::invalid_operation()
+                        .msg("Invalid playlist index.")
+                        .err();
+                };
+                if position > plist.len() {
+                    return Error::invalid_operation()
+                        .msg("Invalid position within playlist.")
+                        .err();
+                }
+                songs.retain(|s| !self.library[s].is_deleted());
+                plist.insert(position, &songs);
+                self.client_update(SubMsg::InsertIntoPlaylist(
+                    InsertIntoPlaylist::new(songs, position, playlist),
+                ));
             }
             IdControlMsg::SetConfig(cfg) => {
                 return Ok(vec![ConfigMsg::Set(cfg).into()]);