+++ src/core/server/sub/mod.rs
@@ -1,4 +1,5 @@
 mod config;
+mod insert_into_playlist;
 mod library;
 mod new_server;
 mod play_tmp;
@@ -12,7 +13,8 @@ mod set_all;
 mod set_playlist;
 
 pub use self::{
-    config::*, library::*, new_server::*, play_tmp::*, player::*, playlist::*,
-    playlist_jump::*, pop_playlist::*, pop_set_playlist::*,
-    reorder_playlist_stack::*, set_all::*, set_playlist::*,
+    config::*, insert_into_playlist::*, library::*, new_server::*,
+    play_tmp::*, player::*, playlist::*, playlist_jump::*, pop_playlist::*,
+    pop_set_playlist::*, reorder_playlist_stack::*, set_all::*,
+    set_playlist::*,
 };